requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
requests
pydantic
pypdf
python-docx
huggingface_hub
